Small One-At-A-Time functions that passes SMHasher
GoodOAAT passes whole SMHasher and has bulk performance greater
than FNV1a (on modern x86).
It could be shortened if avalanche is not an issue (ie if it is
used for hash-tables with prime numbers) in a maner of MicroOAAT.

MicroOAAT - smallest function that passes all collision checks from
SMHasher. It has simpler loop than GoodOAAT, but unfortunately
Clang doesn't produce rotl instruction when `(x>>n|x<<(32-n))` happens
after multiplication, so it is slower than FNV1a with Clang.