harden the short native sha functions
Improved the shortened native SHA functions with seed and len encoded into
the seed, to prevent Zeroes.  These functions need to be padded to 64byte
blocks, so it would be trivial to create max 64 collisions for each key.
Only hardened they would be suitable for common tasks.
The libtomcrypt implementations already encode the original length, so
they are safer against AppendedZeroes.