blake3
with the staticlib from rust (see the PR), and
the C sources. rust not yet.
See the PR https://github.com/BLAKE3-team/BLAKE3/pull/9

Speeds are the same for 30% faster machine with avx2, and the
normal older benchmarking machine with sse41/avx only.

MomentChi2 tests are failing dramatically, even if we mix in
the seed for all 8 IV's