extract Seed_init (info, seed)
and use it for all expensively seeded hashes.

we still pass the seed to the pfhash func, but
some hashes ignore it and need a special Hash_seed_init() for it.
This is for all loops with the same seed, like Hashmap, collision tests
and such. Most tests just use seed 0 btw.
Also requested by hashx, which jit-generates optimized hashfuncs
per seed on the fly. GH #128