Make MomentChi2 results much more human friendly
This aligns printout of the ideal statistical values versus the
observed ones, and aligns the decimals of the chisq values, and places
the chisq values at the end of the line, reducing vertical space and
matching how many other tests work.

This also abstracts out the reporting function, and reorganizes some
comments.