Unify and clarify collision stat reporting
This should not have any major functional changes, it's just making
code clearer and more uniform, reducing redundancy, and making some
comments match the code. It also fixes a bug where '!' strings were
printed on failure even when verbose==false.

The one behavior change is that TestHashList() now calls
PrintCollisions() on failures of hashes of all sizes if drawDiagram is
true. Before it would only be called on hashes with more than 32 bits.