Make pass/warn/fail be based on p-values, and print them out.
This deliberately doesn't print out raw p-values, but instead prints
out the negative log2 of them, which gives a pleasant and compact
order-of-magnitude reporting. Raw p-values are printed in --verbose
mode now.

Deltas are now also displayed in TestBitsCollisions().

Also, all deltas are now only be printed out in --verbose mode,
because they can be too distracting from the ratios, which are almost
always the more important thing for judging hash quality.