Re-add testing for tiny hash table widths, now with useful stats
Since there was a comment from rurban that testing hash behavior on
very small table sizes (8-bit and 12-bit) was desirable, and the
previous testing was not useful, this adds back that testing with some
relevant statistics.

I'm not 100% sold on this addition, and if that is a use-case to be
tested then some other additional approach (like cyan's suggestion of
a statistical test over the populations of all buckets) would well be
useful. But this at least does make the re-added tests' statistical
reports useful and valid.

This also adds a small behavior change where collision subtests are
grouped by bit size instead of high/low bits, since this makes
identical "Expected" numbers appear next to each other.