FIX: Remove compiler warnings in halftime_hash
This stops passing large chunks of data on the stack as function
parameters, using const references instead.

This also aligns some variables that may not have been properly
aligned before, since alignment attributes do not carry through
typedefs OR template parameters. The choice of always aligning to 16
bytes was made as the simplest solution that should always work.

CMake config was updated to silence the warning about the above
attribute issue, now that it has been addressed directly. This is not
a great solution since Hashes.h is _so big_...

Finally, a duplicate CMake line was removed for Hashes.cpp.