Optimize CountMaxCollisions() a bit
This uses window() instead of rshift(), and remembers the result from
the previous trip through the loop to reduce computation.

This also ensures window() is a bit specifically optimized for this
use-case, since every instruction in a long and tight loop matters.

The if() rearrangement is not pretty, but gcc generates better asm
from this format.