Unify collision counting into a single pass using highzerobits()
This is not only a much faster way to compute collision window widths,
but also only needs a single pass over the hashes to compute all
collision counts, both total and maximums.

This also allows for some significant cleanup of TestHashList() and
ReportCollisions(), as well as removal of now-unused collision
counting functions.