Do nbH-appropriate testing for a wider range of bit widths
Since testing additional bit widths is now cheap, thanks to
CountRangedNbCollisions(), this change ensures that testing is done on
as many truncated hash widths as possible, that each of those hash
widths is tested only one way, and that the way each is tested makes
the most mathematical sense.

This also combines 2 calls to CountRangedNbCollisions() into 1,
increasing performance.