Fork smhasher-rurban into smhasher3
First refactor step: split files into major categories, and remove
some files that won't be part of the fork.

This also deletes the crc32c_pclmul hash, since it has strange test
requirements, and that magic hash-specific knowledge gets in the way
of the refactor, and it's in assembly, but mostly because it is
GPL2-only code, which would complicate licensing even further.

Also delete unused Bitslice.cpp