Move MomentChi2 tests into separate .cpp file and rename to Popcount.
The rename was done since that better reflects what is actually being
tested; Moment Chi^2 was just *how* it was being tested.

This test needs some rework, and those magic numbers need to be
documented or *something*. And I have no idea (yet) how to turn the
results into a p-value. :-/