Split out CMakeLists.txt for hashes
This not only shows where the complexities in the file are required,
but also will allow gradual conversion of hashes/ to a new style
without touching the rest of the codebase.

Some minor cleanups are done along the way, most notably introducing
the use of the Cmake CXX_STANDARD variables.