Convert everything to the new HashInfo API
Most of the changes are purely mechanical.

Since each hash implementation will need to be moved individually,
this leaves the old HashInfo API as LegacyHashInfo, and the new API
acts as a wrapper for it. Eventually SMHasher3Hashes and all the
"legacy" code goes away completely.

Some of the new API is "anticipating" some other work I have brewing,
so it does things like just assume a uint64_t type is available.