Remove callback processing from TwoBytes test
The paradigm of providing a callback to a keyset iterator is a cool
one! It not only allows for easily adding new behaviors for the heart
of a keyset generator, it also removes the need for templatizing it
based on the hashtype.

That said, I can't see a real need for iterating over a keyset
multiple times in the same run, which is its largest strength, and all
the other tests don't use it, so I'm deleting it for now, so that all
the tests look much more consistent.

It's very possible this or a similar paradigm will return someday.