Port discohash to new API, and fix many things
A likely-unintentional behavior change was turned into an option (the
"old" and "new" versions of the hash), the 128-bit versions were
added, and the writable hash state was made thread-local to resolve
all the erroneous reports of bad seeds.

Finally, this is now a port of the latest code from
https://github.com/i5ik/discohash, which shows the MIT license instead
of GPL, so that has been updated appropriately.