Port all xxhash hashes to the new API
Also upgraded xxh3 to 0.8.1 release, so those verification codes don't
match the previous SMHasher release. I did verify the new codes
against the reference distribution.

A good chunk of this code is not yet enabled with this commit. A bunch
of instruction set detections (which will enable it) are coming up
next...