Add .impl field for HashInfo
This lets hashes optionally specify a freeform 10-character string describing which
implementation was chosen at compile-time. These strings are printed out in several
places, though often only if --verbose was requested, so as to not clutter the
output.

It is always printed out in the header when a specific hash is tested, if that hash
specified one. It is also always printed out with the --list command.

This fixes issue #61.