Make BIC test use chi-square test of independence
This changes the pass/warn/fail threshholds and the test reporting format both be
more consistent with the rest of SMHasher3. It also changes the --verbose report
slightly. I remain unsure if that report should be scaled by the test count.

This also starts to move analysis and reporting to Analyze.cpp like the rest of
SMHasher3. The existing API is super terrible, and big parts of BicTest4() will need
to be moved over to Analyze.h soon-ish.