util/Stats: Only have a unit normal version of GetNormalPValue()
Since it turns out that's all we ever use, remove the unused code, and
change the function to reflect its new purpose.

Also, the optimizer can evaluate the sqrt() at compile time, but
is forbidden to convert the division to a multiply because
that would change the rounding.

We know it's okay, so express it as a multiply by the reciprocal.