Refactor BIC and SeedBIC test loops
These now match the Avalanche setup, where the same set of key (and seed)
data is used for all tests of a given size, instead of changing it for each
bit being tested.

By itself, this is not a great change for testing fidelity, due to possible
collisions among test data. This will be rectified by a new RNG API shortly.

This change does reduce memory usage significantly, especially for higher
values of --ncpu.