TestDistribution: Improve performance via smaller data arrays
Because the "placing balls into bins" loop accesses the bins randomly,
keeping the bins in L1 cache is critical for performance.

This changes TestDistribution() to start with 8-bit counters, and
degrade to 32-bit counters when needed. The bin array is sized so there
are < 10 hashes per bin on average, so if the hashes are
well-distributed, then the probability of overflowing an 8-bit bin on
initial count should be negligible.

There is additional code complexity detecting and handling overflow
during the folding steps also, but it's still a net win.  The folding
operations access the bin array sequentially, so that's far less
critical.