FilterOutliers: Complete rewrite, now works properly.
The old implementation was simply broken.  Not only did it use a
complicated O(n log n) binary search to perform a fundamentally O(n)
task, but binary search only works if the ContainsOutlier() function
is monotonic, and that is simply false.

Consider a sorted vector of 15x 1.0 and 2x 2.0, and various
prefixes of it:
- Prefix of length <= 15: Standard deviation is 0, no outliers.
- Prefix of length 16: Standard deviation is 0.2421.
  The last element (2.0) is 15/16 = 0.9375 above the mean,
  which is more than 3*sigma = 0.7262.
- Prefix of length 17: Standard deviation is 0.3222.
  The last element (2.0) is 15/17 = 0.8824 above the mean,
  which is less than 3*sigma = 0.9666.

The old code would first check ContainsOutlier(v, 16) and
conclude that more trimming is required, even though
ContainsOutlier(v, 17) is false.

Fortunately, this can all be done with a simple linear search.

Welford's well known, numerically stable algorithm for incrementally
updating the mean and variance of a sample can be used in reverse to
remove a sample in O(1) time.  So we can just start with the
overall variance and repeatedly trim outliers until all is well.

A second change incorporated is the use of the *sample* standard
deviation rather than the *population* deviation.  This slightly
widens the 3*sigma acceptance window, expecially for small number of
samples.