Make p-value handling more consistent
This now treats p-values strictly equal to FAILURE_PBOUND as a failure, and
the same for WARNING_PBOUND. This can fix a very rare issue where a p-value
is reported as ^17 but not be a failure.

This also makes all the p-value plots be consistent and use the same
function and output ASCII scale.