Enable printing of colliding keys
At long last!! So much effort has gone into this feature, and it is finally
working. Now, specifying '--verbose' will print out (a selection of) the
keys that actually collide, as well as their seeds and resulting
hashes. This also handles tests of XORs of hash results.

It's basically impossible to do this for Avalanche or BIC tests, so it
doesn't work for them.